Add files via upload
I managed to do most of the work. some of which was not done was due to timing as I couldn't get the required components, some of which were to be bought. I however attached screenshots of what was achieved. I have little knowledge on mqtt and I was unable do simple test runs with the provided server. I was using the chrome extension version of the mqtt box and attached is the error. Kindly assist with this.